修复二维码生成模块安装判断出错，导致多次安装

+++ user/qrcode.sh
@@ -7,20 +7,22 @@ export PATH=/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin
 #Main
 checkqr(){
 	if [[ ! -e /usr/bin/qr ]];then
-		echo "你还未安装二维码生成模块"
-		echo "按回车键继续，Ctrl+C退出！"
-		read -s
-		echo "正在安装，通常这不需要太多时间"
-		pip -q install qrcode
-		pip -q install git+git://github.com/ojii/pymaging.git
-		pip -q install git+git://github.com/ojii/pymaging-png.git
-		if [[ -e /usr/bin/qr ]];then
-			echo "安装完成！"
-		elif [[ -e /usr/local/bin/qr ]];then
-			echo "安装完成！"
-		else
-			echo "安装失败 请检查你的Python是否正常，并尝试重新安装"
-			exit 1
+		if [[ ! -e /usr/local/bin/qr ]];then
+			echo "你还未安装二维码生成模块"
+			echo "按回车键继续，Ctrl+C退出！"
+			read -s
+			echo "正在安装，通常这不需要太多时间"
+			pip -q install qrcode
+			pip -q install git+git://github.com/ojii/pymaging.git
+			pip -q install git+git://github.com/ojii/pymaging-png.git
+			if [[ -e /usr/bin/qr ]];then
+				echo "安装完成！"
+			elif [[ -e /usr/local/bin/qr ]];then
+				echo "安装完成！"
+			else
+				echo "安装失败 请检查你的Python是否正常，并尝试重新安装"
+				exit 1
+			fi
 		fi
 	fi
 }
